Extended functionality
Added method to allow to iterate over removed items. The idea of this is
to allow disposal of weak needles. Sadly allowing to iterate over the
removed items on a call to Clear means that the bucket cannot be
recycled, since I can't make sure the client will recycle the bucket I
have opted to make the new Clear methods internal.