Reimplanted ArrayPool
The new implementation of ArrayPool prevents creating new works for each
donated array, this way the code can return control faster and generates
less garbage.

If the recycler work was not able to get to a donated array before it is
taken, the new owner automatically performs the corresponding cleaning.

The recycler work happens in the default working context.