Fixes to run tests againts different .NET
ThreadLocal<T>.Values does not exist in .NET 4.0 - but it is backported
from .NET 4.5 to .NET 2.0 up to .NET 3.5.

This means that ThreadLocal<T>.Values works in every .NET except .NET
4.0

If need a solution that works in every .NET use
Theraot.Threading.TrackingThreadLocal<T> - it is a drop in replacement.