Improvements for Weak classes
The problems with Transactions are not fixed

Unit test fail only a few of the times they are run, this makes unit
tests useless to tell me when I introduce bugs, it also makes them
useless to tell me when I have fixed them. Futhermore they help very
little to help me identify the cause of bugs because I cannot reproduce
them.

What I want is to be able to run the unit test in a mode that will
record everything that happens. And run it once and again and again,
keeping only the record of the cases where the unit test fails.
Afterwards I would like to be able to step through the recorded
executions so I can finally gain insight into what went wrong and when.