Recreated Lockables
These clases will use technology developed for Transact (that has been
encapsulated in Pin) to enable (finally!) concurrent data structures.

These classes allow multiple threads to access the needles of a given
context, this context is responsible of serializing the access to the
needles when multiple threads want to access the same needle. There is a
maximun number of threads that can be running on the context. If the
limit is not reached and the threads don't try to access the same
needles at the same time, the threads will not have to wait.