New type: SimpleReentryGuard
The way ReentryGuard used to work was to queue calls and return
IPromise, after the execution finished pending calls in the queue would
be executed, populated the already returned IPromises... This is because
we want every call to ReentryGuard to get executed.

So the idea is that ReentryGuard will do that, and SimpleReentryGuard
will not.