I take onwership of this files
This code was taken from Mono under MIT license, this repository is
under MIT license. If you think that by removing this I'm neglating the
attribution of those who contributed to develop this code, consider two
things: 1) Their names are in the history of the repository to see, 2)
Many others who did contribute has never been named in the file - not
here, nor at mono.