Refactoring ThreadPoolTaskScheduler
The Contract failure happens in the thread that is started for the long
running task. When the Task is passed to Thread.Start it has an Action
when it is recovered in the new thread it doesn't have an Action.

This is not my fault. This seems to be a bug of CLR.