Turns out the ThreadPool is slow - Blame Microsoft!
The default task scheduler is meant to use the ThreadPool, and this
should be visible from the tasks if IsThreadPoolThread is checked, sadly
the queueing work items in the ThreadPool is slow.

Because of that, any test that relies on the result of a race condition
where the default scheduler is expected to be fast... fails. Even tho
they are never observed to fail in modern versions of .NET

I consider relying on the performance of the ThreadPool unreasonable,
and also I have no direct means to improve its performance.

In order to test the hypothesis that the ThreadPool is culprit I did
test with modded schedulers that will inline more aggressively than it,
and also with modded scheduler that add unnecessary delays.

Thanks to that I have marked what tests would fail if executed
sequentially and which are failing because of a slow ThreadPool.

I want to note that some of my own tests resulted in failure when their
execution was serialized (via inlining). These tests work as long as
they run in multiple threads.

There are still a few test marked for review, tho - I'm not sure why
didn't work as expected.

Also I haven't tested with Mono's implementation of the ThreadPool.