Delay "Hack"
Delay promise tasks will be able to check the timeout each time the
Status of the task is checked. This means that if you are checking a
delay in a close loop it may complete even if the internal Timer
responsible for setting the state of task on timeout doesn't run.