Progress...
Some refactorization done
Some clean up done

---

When WaitAllTest gets into the soft lock, the waiting thread is holding
the lock to the tasks continuations it is waiting on, preventing them to
run. I'm unsure how it got the lock.

Futhermore, all taks being waited on claim to have completed, while
WhenAllCore is still expecting the continuation it claims to have added.