Tests Review
Spotted some race conditions in tests. These were hidden because of wait
performance, which is the thing that is being attempted to fix.

Also, there were a lot of cases of not disposed objects in the test...

In fact, these problems play against each other. That is, it is necesary
to make sure that objects gets disposed after the test is done using
them... also, some objects need to be in a particular state before
calling dispose... these things can lead to race conditions.