Gobal Review and Fix Pass 2
Market stuff for futher review. These part of the code might actually be
ok, however I would like to take second look.

Notice: Most of the code maked for review is in the tests. Sadly a lot
of tests around threading and tasks have a comination of race conditions
and disposable objects that can be problematic.