Reworked test
I have learned that I need a race condition.

Why? Well, the thing is that I could write this test without a race
condition, but then a failing condition is that the callback is never
executed which would result in the test never ending.

We do not have time for ethernity, as such I need to set a maximun time
for the callback to run.

However that brings the problem that the test can fail in an old machine
- or even in a modern machine where little CPU time is given to this
proccess.

I honestly do not see a way around. Tests with race conditions are here
to stay.