WhenAny test tweak
This test, as most threading tests relies on race condition.

There are only two options available to me: have race condition or have tests that could run for ever. I am not fighting undecidability, so I am stuck with race conditons.

In this case the test system may attempt to dispose the task before it has finished. So I have added a wait on it to reduce the likelihood of that happening.