Speaking of undecidability...
This test is stuck in the test environment, but it runs correctly on the development environment. I have no idea why, so I am hoping it gets stuck under my debugger.

Well.. hmm... is it still running or is it stuck? I gotta wait a while and comeback and see if it got stuck. Looking at it won't help.