Another wasted test
I hate when this happens, I feel like I am losing my tests.

I am aware that it appears like I am just disabling anything that fails. No, that's not it. I only do this to task that fail because of the timing on race condition.

Yes, I am stuck with that kind of tests. Regardless of the fact that I think these are bad tests, I do not have better tests.

Which is why I hate when this happens... the fact that this test sometimes passes is evidence that it is ok, and the fact that this test sometimes does not pass is evidence that it is sensitive to load.