Double check pattern...
I realized that there is a defect in the test. Given that I call release before writing to the log, there is a chance that the other thread was awaken before this got to write to the log. Resulting in a failed test.

That means that the faster that Release becomes, the greater the chance of a false negative.

This is a fundamental problem with this test... darn you race conditions!

Will see what can I do about it.