Stuck in C# 7.1 - Fixed build
- Language version set to 7.1
- Removed readonly from structs
- KeyCollection and ValueCollection review
- Fixed conditional compilation
- BucketHelper methods with a predicate now end in "Checked" this fixes some ambiguity
- Using auto properties in more situations
- SafeDictionary to never serialize KeyCollection and ValueCollection
- Added JetBrains MSBuild redistributable
- build_msbuild.bat will now use JetBrains MSBuild is MSBuild is not in PATH

WHY? Because the MSBuild from .NET is too old for modern C# and I cannot rely on Visual Studio MSBuild to be available - and if it, it might not be in the same path every time. Thus, I fix it like I fix the Nuget problem: distribute it. However... MSBuild from Microsoft is not redistributable. Thankfully JetBrains did a redistributable one... yet, it only supports up to C# 7.1