Using some name for the thread
Let me explain this to you...

I had only tested that the property Thread.Name would yield the value "threadName". While arguably I cannot test every possible string... I am showing with this that any string could be valid. Consider this test as documentation of how the property behaves (it will not just be "threadName" it will be whatever string I set it to).