.NET Standard 1.0 build
I hope to be able to test all targets from this point on.

This builds on  .NET Standard 1.0, 1.1, 1.2, 1.3 and 1.4

No test run for those targets, yet.

---

As you can see in TypeDiscoverer, we are cheating reflection to make this work. The template must run to get the "reflected" types at compile time.

Note: it is a poor man's parser, and may break.