new TaskEx.FromCancellation
I find it pretty stupid that you can only get a cancelled task from FromCanceled, and that task has nothing to do with the token given as parameter. Furthermore, if I use a token that is not cancelled I get an exception.

If I want a cancelled task - which is the only thing I can get from FromCanceled, aside from the exception - I should NOT need to a token.

Instead, if I am going to use a token, I should get a Task that will be cancelled when the token is. That makes much more sense.