Reworked SemaphoreSlimExtensions.WaitAsync to use ThreadPool
By waiting with ThreadPool.RegisterWaitForSingleObject we do not need to spawn a thread to wait.

The implementation is more complex due to the fact that we need to use SemaphoreSlim.AvailableWaitHandle to be able to use RegisterWaitForSingleObject, and AvailableWaitHandle does not enter the semaphore.

Additional complexity comes from the fact that RegisterWaitForSingleObject does not support CancellationToken