SemaphoreSlimTest file tweak...
I only need Threaot.Core in NET40

The thing is, SemaphoreSlimExtensions only exists in NET40, should I put SemaphoreSlimExtensions in System.Threading? I do not want to, because it won't exist in all targets.

Should I make it exist in all targets and just delegate (like TaskEx)? Perhaps, however, I do not need to do it because this is not the case of being unable to extend a class (like Task) given that I can use extension methods.

It is not really a problem to keep Threaot.Core, the code compiles just fine. The reason why I put it in conditional compilation is not because there is something wrong with the library, but because a tool might remove it considering it not used and break the code.