Re-arranged... most code
I am deciding against Code Maid, and opt for Code Cracker instead.

Code Cracker will allow me to re arrange the whole solution, however it does not handle well conditional compilation that span around methods.

Code Maid will only allow me to re arrange the current file, and will leave it open for me to save it. Which sounds good, but it is a problem for a large solution.

Code Maid does not handle conditional compilation either, what it does is to just ignore it.

Code Cracker tries, and this has resulted in some broken files after it runs... which leads to a new style rule: do not have multiple versions of the same method in different code compilation, use it around partial classes instead.

Code Maid and Code Cracker differ in the order they do things, which means I will have to pick, and I am picking Code Cracker.

The files I have to work on are:
- GCMonitor.Internal
- TypeExtensions
- StreamExtensions
- TaskEx

We can agree they are problematic.