Null reference fix
I really wish that the analytic that tells you to validate parameters on public methods would actually work correctly. It insists I am not validating, forcing me to extract a private method... well, forcing is an strong word. Let us say, compelling. I did not want to disable the analytic because I believe it is useful to be aware of all the places where I might have forgot to validate parameters... then I add the validation and extract a method... however, as evidenced by this fix, I may validate a parameter and forget to validate another.