Nullability review
There is no check for null when creating EnumerableQuery.

Depending on the target you could get a NullReferenceException, a StackOverflow or soft lock when enumerating.

Thus, I do not want to annotate that null is accepted. However, I am not checking for null either. This is not documented to throw ArgumentNullException.

However, I want the code to be aware that it could have null. Even if the code pretends it cannot have null.