NUnit versions for .NET 2.0, 3.0 and .NET 3.5
The good news is that .NET 5.0 SDK still builds .NET 2.0 and up targets. The bad news is that the tests problem is not fixed. The whole problem is that Appveyor is not running NUnit for these targets correctly. I already tried using the newest version. Let us see if an older version known to work solves the issue.